app/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12.4
psycopg2-binary
drf-yasg
